feat: Add custom hook for auth form logic
