+++ src/pages/Register.tsx
@@ -1,41 +1,34 @@
 import * as React from 'react';
-import { useState } from 'react';
 import { useNavigate  } from 'react-router-dom';
 import { useAuthStore } from '../store/authStore';
-import { validateField } from '../utils/validation';
 import { registerUser } from '../api/auth';
+import { useAuthForm } from '../hooks/useAuthForm';
 
 export const Register = () => {
-	const [email, setEmail] = useState<string>('');
-	const [password, setPassword] = useState<string>('');
-	const [emailError, setEmailError] = useState<string | null>(null);
-	const [passwordError, setPasswordError] = useState<string | null>(null);
-	const navigate =useNavigate();
+	const {
+		email,
+		setEmail,
+		password,
+		setPassword,
+		emailError,
+		passwordError,
+		isLoading,
+		setIsLoading,
+		generalError,
+		setGeneralError,
+		validateForm,
+		resetErrors
+	} = useAuthForm();
+
+	const navigate = useNavigate();
 	const login = useAuthStore((state) => state.login);
-	const [isLoading, setIsLoading] = useState<boolean>(false);
-	const [generalError, setGeneralError] = useState<string | null>(null);
 
 	const handleRegister = async (e: React.FormEvent) => {
 		e.preventDefault();
 
-		// Використовуємо функцію валідації
-		const newEmailError = validateField({ type: 'email', value: email, required: true } as HTMLInputElement);
-		const newPasswordError = validateField({ type: 'password', value: password, required: true } as HTMLInputElement);
-
-		// Оновлюємо стан помилок
-		setEmailError(newEmailError);
-		setPasswordError(newPasswordError);
-
-		// Перевіряємо, чи немає помилок валідації
-		if (newEmailError || newPasswordError) {
-			// Якщо є помилки, зупиняємо виконання функції
+		if (!validateForm) {
 			return;
 		}
-
-		// const newUser = {
-		//   id: Date.now().toString(),
-		// 	email,
-		// }
 
 		setIsLoading(true); // Включаємо стан завантаження
 		setGeneralError(null); // Очищаємо попередні помилки
@@ -69,9 +62,9 @@ export const Register = () => {
 				id="email"
 				type="email"
 				value={email}
-				onChange={(e:React.ChangeEvent<HTMLInputElement>) => {
+				onChange={(e) => {
 					setEmail(e.target.value);
-					setGeneralError(null);
+					resetErrors(null);
 				}}
 				placeholder="Email"
 				required
@@ -83,9 +76,9 @@ export const Register = () => {
 				id="password"
 				type="password"
 				value={password}
-				onChange={(e:React.ChangeEvent<HTMLInputElement>) => {
+				onChange={(e) => {
 					setPassword(e.target.value);
-					setPasswordError(null);
+					resetErrors(null);
 				}}
 				placeholder="Password"
 				required
